refactor gomod module logic
