+++ tools/goctl/api/gogen/util.go
@@ -4,41 +4,21 @@ import (
 	"fmt"
 	goformat "go/format"
 	"io"
-	"os"
-	"path"
-	"path/filepath"
 	"strings"
 
 	"github.com/tal-tech/go-zero/core/collection"
 	"github.com/tal-tech/go-zero/tools/goctl/api/spec"
 	"github.com/tal-tech/go-zero/tools/goctl/api/util"
-	goctlutil "github.com/tal-tech/go-zero/tools/goctl/util"
+	"github.com/tal-tech/go-zero/tools/goctl/util/project"
 )
 
 func getParentPackage(dir string) (string, error) {
-	absDir, err := filepath.Abs(dir)
+	p, err := project.Prepare(dir, false)
 	if err != nil {
 		return "", err
 	}
 
-	absDir = strings.ReplaceAll(absDir, `\`, `/`)
-	rootPath, hasGoMod := goctlutil.FindGoModPath(dir)
-	if hasGoMod {
-		return rootPath, nil
-	}
-
-	gopath := os.Getenv("GOPATH")
-	parent := path.Join(gopath, "src")
-	pos := strings.Index(absDir, parent)
-	if pos < 0 {
-		fmt.Printf("%s not in go.mod project path, or not in GOPATH of %s directory\n", absDir, gopath)
-		tempPath := filepath.Dir(absDir)
-		rootPath = absDir[len(tempPath)+1:]
-	} else {
-		rootPath = absDir[len(parent)+1:]
-	}
-
-	return rootPath, nil
+	return p.GoMod.Module, nil
 }
 
 func writeIndent(writer io.Writer, indent int) {
